@@ -30,7 +30,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.2",
+        "php": ">=5.4",
         "omnipay/common": "~2.3",
         "moip/moip-sdk-php": "~3.1"
     },
